+++ CHANGES.md
@@ -20,6 +20,7 @@
 * Change enable image caching on browse pages
 * Change update sceneNameCache after scene names are updated
 * Change add core dedicated base class tvinfo_base to unify future info sources
+* Add exclude ignore words and exclude required words to settings/Search, Edit and View show
 
 
 [develop changelog]